pyproject.toml
[project]
name = "siem-assistant-mvp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datetime>=5.5",
    "elasticsearch>=9.1.1",
    "google-generativeai>=0.8.5",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "streamlit>=1.49.1",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "pytest>=8.4.2",
]
